backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy>=2.0.35,<2.1
pydantic==2.9.2
python-dotenv
sqlalchemy-libsql>=0.2.0
# opcional, geralmente vem como dependência do sqlalchemy-libsql:
libsql-experimental>=0.0.55
